// src/api-client/models/partner-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PartnershipStatus } from './partnership-status';
// May contain unused imports in some cases
// @ts-ignore
import type { RecipientType } from './recipient-type';

/**
 * 
 * @export
 * @interface PartnerResponseDto
 */
export interface PartnerResponseDto {
    /**
     * The partnership row.
     * @type {string}
     * @memberof PartnerResponseDto
     */
    'partnershipId': string;
    /**
     * The recipient.
     * @type {string}
     * @memberof PartnerResponseDto
     */
    'recipientId': string;
    /**
     * Public name.
     * @type {string}
     * @memberof PartnerResponseDto
     */
    'displayName': string;
    /**
     * Abbreviated name, for tight layouts.
     * @type {string}
     * @memberof PartnerResponseDto
     */
    'shortName'?: string | null;
    /**
     * What kind of receiver they are.
     * @type {RecipientType}
     * @memberof PartnerResponseDto
     */
    'type': RecipientType;
    /**
     * Logo URL.
     * @type {string}
     * @memberof PartnerResponseDto
     */
    'logoUrl'?: string | null;
    /**
     * Whether an admin has verified them.
     * @type {boolean}
     * @memberof PartnerResponseDto
     */
    'isVerified': boolean;
    /**
     * State of the partnership.
     * @type {PartnershipStatus}
     * @memberof PartnerResponseDto
     */
    'partnershipStatus': PartnershipStatus;
    /**
     * Whether this partner is preferred.
     * @type {boolean}
     * @memberof PartnerResponseDto
     */
    'isPreferred': boolean;
    /**
     * Named contact on their side.
     * @type {string}
     * @memberof PartnerResponseDto
     */
    'contactPerson'?: string | null;
    /**
     * Their phone.
     * @type {string}
     * @memberof PartnerResponseDto
     */
    'phone'?: string | null;
    /**
     * Plate of their usual vehicle.
     * @type {string}
     * @memberof PartnerResponseDto
     */
    'vehiclePlate'?: string | null;
    /**
     * Donations delivered to them from this branch.
     * @type {number}
     * @memberof PartnerResponseDto
     */
    'deliveredCount': number;
}