// src/api-client/models/partnership-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * State of the relationship.
 * @export
 * @enum {string}
 */

export const PartnershipStatus = {
    Pending: 'PENDING',
    Active: 'ACTIVE',
    Paused: 'PAUSED',
    Ended: 'ENDED'
} as const;

export type PartnershipStatus = typeof PartnershipStatus[keyof typeof PartnershipStatus];